backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
pydantic[email]==2.9.2
python-dotenv==1.0.1
httpx[http2]==0.27.2
